[TM-47] fix: Fixed all comented flaws
