+++ src/utils/constants/index.ts
@@ -1,7 +1,6 @@
 import { CoverType } from '../../domains/client';
 
 export const EMAIL_BASE_STRUCTURE = /^(?=.{1,254}$)[^\s@]+@[^\s@]+\.[^\s@]+$/;
-
 export const CODE_BASE_STRUCTURE = /^\d{6}$/;
 
 export const ALG = 'sha256';
@@ -9,17 +8,12 @@ export const ALG = 'sha256';
 export const CODE_LENGTH = 6;
 
 export const OTP_CODE_MIN_DIGIT = 0;
-
 export const OTP_CODE_MAX_DIGIT = 10;
-
 export const OTP_SALT_LENGTH = 16;
-
 export const OTP_TTL_MS = 60 * 60 * 1000;
 
 export const SMTP_MAX_SEND_ATTEMPTS = 2;
-
 export const SMTP_MIN_SEND_ATTEMPTS = 1;
-
 export const SMTP_RETRY_DELAY_MS = 300;
 
 export const BARCODE_KEYS = ['ean13', 'gtin', 'ean8', 'code128'] as const;
@@ -64,17 +58,11 @@ export const MOYSKLAD_HOSTNAME = (() => {
 })();
 
 export const DEFAULT_MOYSKLAD_HOSTNAME = 'api.moysklad.ru';
-
 export const RE_DOWNLOAD = /\/download\/[0-9a-f-]{36}$/i;
-
 export const RE_ENTITY_IMAGE = /\/entity\/image\/[0-9a-f-]{36}$/i;
 
 export const PRODUCT_RATING_MIN = 1;
-
 export const PRODUCT_RATING_MAX = 4;
-
 export const DEFAULT_REVIEWS_LIMIT = 3;
-
 export const DEFAULT_NUMERIC_VALUE = 0;
-
 export const ANONYMOUS_AUTHOR_FALLBACK = 'Аноним';
